pyproject.toml
[tool.poetry]
name = "lending-library"
version = "0.1.0"
description = "A flask api for a book lending library"
authors = ["Michael Lane <mikelane@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-RESTful = "^0.3.8"
Flask-SQLAlchemy = "^2.4.4"
alembic = "^1.5.5"
SQLAlchemy-Searchable = "^1.2.0"
flask-marshmallow = "^0.14.0"
Flask-Admin = "^1.5.7"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.0.4"
python-dotenv = "^0.15.0"
marshmallow-sqlalchemy = "^0.24.2"
SQLAlchemy-Utils = "^0.36.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-console-scripts = "^1.1.0"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
pytest-picked = "^0.4.6"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.2.1"
black = {version = "^20.8b1", allow-prereleases = true}
pytest-black = {version = "^0.3.12", allow-prereleases = true}
isort = {version = "^5.7.0", allow-prereleases = true}
pytest-isort = {version = "^1.3.0", allow-prereleases = true}
mypy = {version = "^0.812", allow-prereleases = true}
pytest-mypy = {version = "^0.8.0", allow-prereleases = true}
pytest-common-subject = "^1.0.5"

[tool.black]
line-length = 110
include = ".pyi?$"
exclude = '''
(
  (
    \.eggs 
    | \.git 
    | \.mypy_cache
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
